There seemed to be some confusion between using type '#' and type 'placeholder in the code. I switched everything to be 'placeholder'.
The real fix is that in Parser.prototype.scanToken, there was one 'if'
testing for '#' and another testing for 'placeholder'. The fix was to move
the 'placeholder' test into the '#' test so that when there was not
something to replace, a placeholder was inserted into the result.

Note: the previous code in that case (which probably was never used) used
the value of the placeholder token (typically a '0') for the placeholder.
That's not great visually, so I switch to using 'U+2753 = BLACK QUESTION
MARK ORNAMENT' which was used elsewhere.

Fixes #33.