More improvements for accessibility. One big one is changing the textedit to have aria-live="hidden" tabindex="0". This allows focus to get to the right area when tabbing but not have screen readers echo arrow key and other "special" chars.
Added a bunch more cases for speaking.

The onfocus speech is good to do in some cases and bad in others (particular after a paste from a palette). Still thinking about it.

This includes a change to where the caret is placed when the numerator is deleted. It is now before the contents of what was the denominator since that is where it should be in the typing and reading order.