fix: properly observe `read-only` changes
Before, changes to the `readonly` attribute were observed but changes to `read-only` were not. This was causing wrong behaviors when adding `read-only` to a `math-field` element, for example it was allowing users to modify the content of a `math-field` after adding the `read-only` attribute programmatically.
You can see here the behavior of a `math-field` after toggling `read-only` or `readonly`:
https://codepen.io/LuisMesa/pen/GROgOZz?editors=1010