fix: Firefox and Safari selection issues, fixes #1898
Adjusted the logic of when keyboardSink selection buffer is cleared. It is now cleared on keydown for any input that changes the buffer (fixes Firefox issue where it wasn't possible to replace a selection with a key press that wasn't a shortcut).

Buffer is not automatically cleared for non content changing command so that Cmd-A and Shift-Arrow work to select on Safari. Arrow keys alone still work to clear selection buffer since this is handled by the current selection update logic.