fix: Deleting the Entire Math Expression by Backspacing on a Specific Symbol
Reproducible Procedure:

1. Prepare a math-field and input the expression \sqrt{a}b.
2. Select the symbol a that is inside the square root and press the
backspace key.
3. The entire expression \sqrt{a}b will be deleted, leaving the input
field blank.

Additional Information:

In this implementation, when deleting a child element such as "a" that is inside the square root the logic is designed to delete the entire expression. However, the
deletion range is not properly defined, which causes the deletion of
an unintended element one position to the right of the parent element.