fix: Safari context menu cut fix (#2326)
* fix: Safari context menu cut fix

Fixes context menu cut functionality for Safari (previously cut would do nothing on Safari). Tests have also been added for both keyboard cut  and paste shortcuts and also for context menu cut and paste.

* tests: Fix failing cut/paste tests

In CI on firefox, insignificant spaces are removed in comparison. For webkit, tests that use keyboard paste are skipped in CI since the keyboard paste doesn't work for webkit when headless on Linux. Keyboard paste does work when headed on Linux or headless/headed on MacOs.

* fix: apply same fix to copyToClipboard as was applied to cutToClipboard

Issue only occurs when using the `copyToClipboard` command on Safari when there is a active selection, nothing was selected in this case.